fix #34 - alternative way to group services
Reason for this change is that using SvcHostSplitThresholdInKB (FFFFFFFF) breaks XboxGipSvc, and possibly others that we are not aware of yet.

The alternative way is to identify (most) grouped services and add a SvcHostSplitDisable for each service. This should yield similar results to setting SvcHostSplitThresholdInKB. The list included in SVCGROUP.ps1 might not be final.