add writeStatus for each action in appx.yml
When attempting to remove certain appx packages, there may be cases where the package does not get deleted on the first try due to locks. In such cases, the AME Wizard's GUI status reporter may mistakenly perceive an issue when there isn't one, causing "Reticulating Splines" to appear until the next "writeStatus". To address this issue, adding a status for each action in appx.yml seems to eliminate it.